Fix: 게시글 공감 추가 및 삭제 예외 처리, 응답 값 수정, prefix 수정

+++ src/main/java/com/cocos/cocos/api/post/service/PostLikeService.java
@@ -1,7 +1,9 @@
 package com.cocos.cocos.api.post.service;
 
+import com.cocos.cocos.common.exception.CocosException;
 import com.cocos.cocos.db.post.entity.PostLike;
 import com.cocos.cocos.db.post.repository.PostLikeRepository;
+import com.cocos.cocos.enums.message.FailMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +16,9 @@ public class PostLikeService {
 
     @Transactional
     public void addPostLike(final Long memberId, final Long postId) {
+        if (postLikeRepository.existsByMemberIdAndPostId(memberId, postId)) {
+            throw new CocosException(FailMessage.CONFLICT_POSTLIKE);
+        }
         postLikeRepository.save(
                 PostLike.builder()
                         .memberId(memberId)
@@ -24,6 +29,9 @@ public class PostLikeService {
 
     @Transactional
     public void deletePostLike(final Long memberId, final Long postId) {
+        if (!postLikeRepository.existsByMemberIdAndPostId(memberId, postId)) {
+            throw new CocosException(FailMessage.NOT_FOUND_POSTLIKE);
+        }
         postLikeRepository.deleteByMemberIdAndPostId(memberId, postId);
     }
 }
